Can't delete job with id of None (#42)
* Avoid running "qdel" with a null job number

* Add file size to debug log

- Display file size of uploaded and downloaded files in debug output.
- Add another exception type to debug logging to get file size to work well.
- Remove unicode re-encoding. Adding 'encoding: utf-8' to the file logger config is the proper fix.
- Remove unnecessary debug_start_time from get_file() debug output.

* Avoid error with empty list

This error was highlighted in a database migration from
https://github.com/erdc/django-uit_job_type/pull/16

+++ uit/job.py
@@ -164,7 +164,12 @@ class PbsJob:
         )
 
     def terminate(self):
-        return self._execute('qdel')
+        if self.job_id:
+            return self._execute('qdel')
+        else:
+            # Avoid running "qdel None" which causes an error. There is nothing to terminate in PBS, so return True.
+            # This is often the result of a job that created files on the HPC but failed to create the PBS job.
+            return True
 
     def hold(self):
         return self._execute('qhold')
